requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.16.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
Django==4.2.11
flask==3.0.2
geojson==3.1.0
idna==3.6
image==1.5.33
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nest-asyncio==1.6.0
novita-client==0.5.3
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pillow==10.3.0
plotly==5.19.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
retrying==1.3.4
six==1.16.0
sqlparse==0.5.0
tenacity==8.2.3
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
